@@ -1,29 +1,32 @@
 {
-  "name": "vue",
+  "name": "nkdayscraper",
   "version": "0.1.0",
   "private": true,
   "scripts": {
+    "serve-prod": "vue-cli-service serve",
+    "build-prod": "vue-cli-service build",
+    "lint-prod": "vue-cli-service lint",
     "serve": "vue-cli-service serve --mode development",
     "build": "vue-cli-service build --mode development",
     "lint": "vue-cli-service lint --mode development"
   },
   "dependencies": {
-    "@vue/composition-api": "^1.0.0-beta.6",
     "axios": "^0.19.2",
-    "bootstrap-vue": "^2.15.0",
+    "bootstrap-vue": "^2.16.0",
     "core-js": "^3.6.5",
     "jquery": "^3.5.1",
     "typescript": "^3.9.7",
-    "vue": "^2.6.11"
+    "vue": "^3.0.0-rc.5"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.4.0",
     "@vue/cli-plugin-eslint": "~4.4.0",
     "@vue/cli-service": "~4.4.0",
+    "@vue/compiler-sfc": "^3.0.0-rc.5",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.11"
+    "eslint-plugin-vue": "^7.0.0-beta.1",
+    "vue-cli-plugin-vue-next": "~0.1.3"
   },
   "eslintConfig": {
     "root": true,
@@ -31,7 +34,7 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
     "parserOptions": {
