@@ -1,12 +1,7 @@
-import Vue from 'vue'
-import VueCompositionAPI from '@vue/composition-api'
+import { createApp } from 'vue';
+// import VueCompositionAPI from 'vue'
 import App from './App.vue'
 import 'bootstrap/dist/css/bootstrap.css'
 import 'bootstrap-vue/dist/bootstrap-vue.css'
 
-Vue.config.productionTip = false
-Vue.use(VueCompositionAPI)
-
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+createApp(App).mount('#app')
